+++ internal/observability/call.go
@@ -83,6 +83,8 @@ func (c call) EndWithPanic(err error) {
 }
 
 func (c call) endLogs(elapsed time.Duration, err error, isApplicationError bool) {
+	isThriftException := isApplicationError && err == nil
+
 	var ce *zapcore.CheckedEntry
 	if err == nil && !isApplicationError {
 		msg := _successfulInbound
@@ -97,7 +99,16 @@ func (c call) endLogs(elapsed time.Duration, err error, isApplicationError bool)
 		}
 
 		lvl := c.levels.failure
-		if isApplicationError {
+
+		// Application errors are
+		//  - Thrift exceptions
+		//  - `yarpcerror`s with error details
+		//
+		// Any error returned from a Protobuf handler is marked as an application
+		// error (unlike Thrift). Therefore, we distinguish an application error
+		// from a regular error by inspecting if an error detail was set.
+		isProtoErrDetail := isApplicationError && len(yarpcerrors.FromError(err).Details()) > 0
+		if isThriftException || isProtoErrDetail {
 			lvl = c.levels.applicationError
 		}
 		ce = c.edge.logger.Check(lvl, msg)
@@ -112,7 +123,7 @@ func (c call) endLogs(elapsed time.Duration, err error, isApplicationError bool)
 	fields = append(fields, zap.Duration("latency", elapsed))
 	fields = append(fields, zap.Bool("successful", err == nil && !isApplicationError))
 	fields = append(fields, c.extract(c.ctx))
-	if isApplicationError {
+	if isThriftException {
 		fields = append(fields, zap.String(_error, "application_error"))
 	} else {
 		fields = append(fields, zap.Error(err))